+++ tests/subsys/rtio/rtio_api/CMakeLists.txt
@@ -5,7 +5,7 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(rtio_api_test)
 
-target_sources(app PRIVATE src/main.c)
+target_sources(app PRIVATE src/test_rtio_spsc.c src/test_rtio_mpsc.c src/test_rtio_api.c)
 
 target_include_directories(app PRIVATE
   ${ZEPHYR_BASE}/include
