+++ tests/subsys/rtio/rtio_api/src/test_rtio_api.c
@@ -1,0 +1,293 @@
+/*
+ * Copyright (c) 2021 Intel Corporation.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <errno.h>
+#include <zephyr/ztest.h>
+#include <zephyr/kernel.h>
+#include <zephyr/sys/atomic.h>
+#include <zephyr/sys/kobject.h>
+#include <zephyr/sys/libc-hooks.h>
+#include <zephyr/app_memory/mem_domain.h>
+#include <zephyr/sys/util_loops.h>
+#include <zephyr/sys/time_units.h>
+#include <zephyr/rtio/rtio_spsc.h>
+#include <zephyr/rtio/rtio_mpsc.h>
+#include <zephyr/rtio/rtio.h>
+#include <zephyr/rtio/rtio_executor_simple.h>
+#include <zephyr/rtio/rtio_executor_concurrent.h>
+
+#include "rtio_iodev_test.h"
+
+RTIO_EXECUTOR_SIMPLE_DEFINE(simple_exec_simp);
+RTIO_DEFINE(r_simple_simp, (struct rtio_executor *)&simple_exec_simp, 4, 4);
+
+RTIO_EXECUTOR_CONCURRENT_DEFINE(simple_exec_con, 1);
+RTIO_DEFINE(r_simple_con, (struct rtio_executor *)&simple_exec_con, 4, 4);
+
+RTIO_IODEV_TEST_DEFINE(iodev_test_simple);
+
+/**
+ * @brief Test the basics of the RTIO API
+ *
+ * Ensures that we can setup an RTIO context, enqueue a request, and receive
+ * a completion event.
+ */
+void test_rtio_simple_(struct rtio *r)
+{
+	int res;
+	uintptr_t userdata[2] = {0, 1};
+	struct rtio_sqe *sqe;
+	struct rtio_cqe *cqe;
+
+	rtio_iodev_test_init(&iodev_test_simple);
+
+	TC_PRINT("setting up single no-op\n");
+	sqe = rtio_spsc_acquire(r->sq);
+	zassert_not_null(sqe, "Expected a valid sqe");
+	rtio_sqe_prep_nop(sqe, (struct rtio_iodev *)&iodev_test_simple, &userdata[0]);
+
+	TC_PRINT("submit with wait\n");
+	res = rtio_submit(r, 1);
+	zassert_ok(res, "Should return ok from rtio_execute");
+
+	cqe = rtio_spsc_consume(r->cq);
+	zassert_not_null(cqe, "Expected a valid cqe");
+	zassert_ok(cqe->result, "Result should be ok");
+	zassert_equal_ptr(cqe->userdata, &userdata[0], "Expected userdata back");
+	rtio_spsc_release(r->cq);
+}
+
+ZTEST(rtio_api, test_rtio_simple)
+{
+	TC_PRINT("rtio simple simple\n");
+	test_rtio_simple_(&r_simple_simp);
+	TC_PRINT("rtio simple concurrent\n");
+	test_rtio_simple_(&r_simple_con);
+}
+
+RTIO_EXECUTOR_SIMPLE_DEFINE(chain_exec_simp);
+RTIO_DEFINE(r_chain_simp, (struct rtio_executor *)&chain_exec_simp, 4, 4);
+
+RTIO_EXECUTOR_CONCURRENT_DEFINE(chain_exec_con, 1);
+RTIO_DEFINE(r_chain_con, (struct rtio_executor *)&chain_exec_con, 4, 4);
+
+RTIO_IODEV_TEST_DEFINE(iodev_test_chain0);
+RTIO_IODEV_TEST_DEFINE(iodev_test_chain1);
+struct rtio_iodev *iodev_test_chain[] = {&iodev_test_chain0, &iodev_test_chain1};
+
+/**
+ * @brief Test chained requests
+ *
+ * Ensures that we can setup an RTIO context, enqueue a chained requests,
+ * and receive completion events in the correct order given the chained
+ * flag and multiple devices where serialization isn't guaranteed.
+ */
+void test_rtio_chain_(struct rtio *r)
+{
+	int res;
+	uintptr_t userdata[4] = {0, 1, 2, 3};
+	struct rtio_sqe *sqe;
+	struct rtio_cqe *cqe;
+
+	for (int i = 0; i < 4; i++) {
+		sqe = rtio_spsc_acquire(r->sq);
+		zassert_not_null(sqe, "Expected a valid sqe");
+		rtio_sqe_prep_nop(sqe, iodev_test_chain[i % 2],
+				  &userdata[i]);
+		sqe->flags |= RTIO_SQE_CHAINED;
+	}
+
+	/* Clear the last one */
+	sqe->flags = 0;
+
+	TC_PRINT("submitting\n");
+	res = rtio_submit(r, 4);
+	TC_PRINT("checking cq\n");
+	zassert_ok(res, "Should return ok from rtio_execute");
+	zassert_equal(rtio_spsc_consumable(r->cq), 4, "Should have 4 pending completions");
+
+	for (int i = 0; i < 4; i++) {
+		TC_PRINT("consume %d\n", i);
+		cqe = rtio_spsc_consume(r->cq);
+		zassert_not_null(cqe, "Expected a valid cqe");
+		zassert_ok(cqe->result, "Result should be ok");
+		zassert_equal_ptr(cqe->userdata, &userdata[i], "Expected in order completions");
+		rtio_spsc_release(r->cq);
+	}
+}
+
+ZTEST(rtio_api, test_rtio_chain)
+{
+	TC_PRINT("initializing iodev test devices\n");
+
+	for (int i = 0; i < 2; i++) {
+		rtio_iodev_test_init(iodev_test_chain[i]);
+	}
+
+	TC_PRINT("rtio chain simple\n");
+	test_rtio_chain_(&r_chain_simp);
+	TC_PRINT("rtio chain concurrent\n");
+	test_rtio_chain_(&r_chain_con);
+}
+
+
+RTIO_EXECUTOR_SIMPLE_DEFINE(multi_exec_simp);
+RTIO_DEFINE(r_multi_simp, (struct rtio_executor *)&multi_exec_simp, 4, 4);
+
+RTIO_EXECUTOR_CONCURRENT_DEFINE(multi_exec_con, 2);
+RTIO_DEFINE(r_multi_con, (struct rtio_executor *)&multi_exec_con, 4, 4);
+
+RTIO_IODEV_TEST_DEFINE(iodev_test_multi0);
+RTIO_IODEV_TEST_DEFINE(iodev_test_multi1);
+struct rtio_iodev *iodev_test_multi[] = {&iodev_test_multi0, &iodev_test_multi1};
+
+/**
+ * @brief Test multiple asynchronous chains against one iodev
+ */
+void test_rtio_multiple_chains_(struct rtio *r)
+{
+	int res;
+	uintptr_t userdata[4] = {0, 1, 2, 3};
+	struct rtio_sqe *sqe;
+	struct rtio_cqe *cqe;
+
+	for (int i = 0; i < 2; i++) {
+		for (int j = 0; j < 2; j++) {
+			sqe = rtio_spsc_acquire(r->sq);
+			zassert_not_null(sqe, "Expected a valid sqe");
+			rtio_sqe_prep_nop(sqe, iodev_test_multi[i],
+					  (void *)userdata[i*2 + j]);
+			if (j == 0) {
+				sqe->flags |= RTIO_SQE_CHAINED;
+			} else {
+				sqe->flags |= 0;
+			}
+		}
+	}
+
+	TC_PRINT("calling submit from test case\n");
+	res = rtio_submit(r, 0);
+	zassert_ok(res, "Should return ok from rtio_execute");
+
+	bool seen[4] = { 0 };
+
+	TC_PRINT("waiting for 4 completions\n");
+	for (int i = 0; i < 4; i++) {
+		TC_PRINT("waiting on completion %d\n", i);
+		cqe = rtio_spsc_consume(r->cq);
+
+		while (cqe == NULL) {
+			k_sleep(K_MSEC(1));
+			cqe = rtio_spsc_consume(r->cq);
+		}
+
+		zassert_not_null(cqe, "Expected a valid cqe");
+		TC_PRINT("result %d, would block is %d, inval is %d\n",
+			 cqe->result, -EWOULDBLOCK, -EINVAL);
+		zassert_ok(cqe->result, "Result should be ok");
+		seen[(uintptr_t)cqe->userdata] = true;
+		if (seen[1]) {
+			zassert_true(seen[0], "Should see 0 before 1");
+		}
+		if (seen[3]) {
+			zassert_true(seen[2], "Should see 2 before 3");
+		}
+		rtio_spsc_release(r->cq);
+	}
+}
+
+ZTEST(rtio_api, test_rtio_multiple_chains)
+{
+	for (int i = 0; i < 2; i++) {
+		rtio_iodev_test_init(iodev_test_multi[i]);
+	}
+
+	TC_PRINT("rtio multiple simple\n");
+	test_rtio_multiple_chains_(&r_multi_simp);
+	TC_PRINT("rtio_multiple concurrent\n");
+	test_rtio_multiple_chains_(&r_multi_con);
+}
+
+
+
+#ifdef CONFIG_USERSPACE
+K_APPMEM_PARTITION_DEFINE(rtio_partition);
+K_APP_BMEM(rtio_partition) uint8_t syscall_bufs[4];
+struct k_mem_domain rtio_domain;
+#else
+uint8_t syscall_bufs[4];
+#endif
+
+RTIO_EXECUTOR_SIMPLE_DEFINE(syscall_simple);
+RTIO_DEFINE(r_syscall, (struct rtio_executor *)&syscall_simple, 4, 4);
+RTIO_IODEV_TEST_DEFINE(iodev_test_syscall);
+
+void rtio_syscall_test(void *p1, void *p2, void *p3)
+{
+	int res;
+	struct rtio_sqe sqe;
+	struct rtio_cqe cqe;
+
+	struct rtio *r = &r_syscall;
+
+	for (int i = 0; i < 4; i++) {
+		TC_PRINT("copying sqe in from stack\n");
+		/* Not really legal from userspace! Ugh */
+		rtio_sqe_prep_nop(&sqe, &iodev_test_syscall,
+				  &syscall_bufs[i]);
+		res = rtio_sqe_copy_in(r, &sqe, 1);
+		zassert_equal(res, 0, "Expected success copying sqe");
+	}
+
+	TC_PRINT("submitting\n");
+	res = rtio_submit(r, 4);
+
+	for (int i = 0; i < 4; i++) {
+		TC_PRINT("consume %d\n", i);
+		res = rtio_cqe_copy_out(r, &cqe, 1, K_FOREVER);
+		zassert_equal(res, 1, "Expected success copying cqe");
+		zassert_ok(cqe.result, "Result should be ok");
+		zassert_equal_ptr(cqe.userdata, &syscall_bufs[i],
+				  "Expected in order completions");
+	}
+}
+
+#ifdef CONFIG_USERSPACE
+ZTEST(rtio_api, test_rtio_syscalls_usermode)
+{
+	struct k_mem_partition *parts[] = {
+#if Z_LIBC_PARTITION_EXISTS
+		&z_libc_partition,
+#endif
+		&rtio_partition
+	};
+
+	TC_PRINT("syscalls from user mode test\n");
+	TC_PRINT("test iodev init\n");
+	rtio_iodev_test_init(&iodev_test_syscall);
+	TC_PRINT("mem domain init\n");
+	k_mem_domain_init(&rtio_domain, ARRAY_SIZE(parts), parts);
+	TC_PRINT("mem domain add current\n");
+	k_mem_domain_add_thread(&rtio_domain, k_current_get());
+	TC_PRINT("rtio access grant\n");
+	rtio_access_grant(&r_syscall, k_current_get());
+	TC_PRINT("rtio iodev access grant, ptr %p\n", &iodev_test_syscall);
+	k_object_access_grant(&iodev_test_syscall, k_current_get());
+	TC_PRINT("user mode enter\n");
+	k_thread_user_mode_enter(rtio_syscall_test, NULL, NULL, NULL);
+}
+#endif /* CONFIG_USERSPACE */
+
+
+ZTEST(rtio_api, test_rtio_syscalls)
+{
+	TC_PRINT("test iodev init\n");
+	rtio_iodev_test_init(&iodev_test_syscall);
+	TC_PRINT("syscalls from kernel mode\n");
+	rtio_syscall_test(NULL, NULL, NULL);
+}
+
+ZTEST_SUITE(rtio_api, NULL, NULL, NULL, NULL, NULL);
