+++ tests/subsys/rtio/rtio_api/src/test_rtio_mpsc.c
@@ -1,0 +1,179 @@
+/*
+ * Copyright (c) 2023 Intel Corporation.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <zephyr/ztest.h>
+#include <zephyr/kernel.h>
+#include <zephyr/sys/util_loops.h>
+#include <zephyr/rtio/rtio_spsc.h>
+#include <zephyr/rtio/rtio_mpsc.h>
+
+#include "rtio_api.h"
+
+static struct rtio_mpsc push_pop_q;
+static struct rtio_mpsc_node push_pop_nodes[2];
+
+/*
+ * @brief Push and pop one element
+ *
+ * @see rtio_mpsc_push(), rtio_mpsc_pop()
+ *
+ * @ingroup rtio_tests
+ */
+ZTEST(rtio_mpsc, test_push_pop)
+{
+
+	struct rtio_mpsc_node *node, *head, *stub, *next, *tail;
+
+	rtio_mpsc_init(&push_pop_q);
+
+	head = atomic_ptr_get(&push_pop_q.head);
+	tail = push_pop_q.tail;
+	stub = &push_pop_q.stub;
+	next = atomic_ptr_get(&stub->next);
+
+	zassert_equal(head, stub, "Head should point at stub");
+	zassert_equal(tail, stub, "Tail should point at stub");
+	zassert_is_null(next, "Next should be null");
+
+	node = rtio_mpsc_pop(&push_pop_q);
+	zassert_is_null(node, "Pop on empty queue should return null");
+
+	rtio_mpsc_push(&push_pop_q, &push_pop_nodes[0]);
+
+	head = atomic_ptr_get(&push_pop_q.head);
+
+	zassert_equal(head, &push_pop_nodes[0], "Queue head should point at push_pop_node");
+	next = atomic_ptr_get(&push_pop_nodes[0].next);
+	zassert_is_null(next, NULL, "push_pop_node next should point at null");
+	next = atomic_ptr_get(&push_pop_q.stub.next);
+	zassert_equal(next, &push_pop_nodes[0], "Queue stub should point at push_pop_node");
+	tail = push_pop_q.tail;
+	stub = &push_pop_q.stub;
+	zassert_equal(tail, stub, "Tail should point at stub");
+
+	node = rtio_mpsc_pop(&push_pop_q);
+	stub = &push_pop_q.stub;
+
+	zassert_not_equal(node, stub, "Pop should not return stub");
+	zassert_not_null(node, "Pop should not return null");
+	zassert_equal(node, &push_pop_nodes[0],
+		      "Pop should return push_pop_node %p, instead was %p",
+		      &push_pop_nodes[0], node);
+
+	node = rtio_mpsc_pop(&push_pop_q);
+	zassert_is_null(node, "Pop on empty queue should return null");
+}
+
+#define MPSC_FREEQ_SZ 8
+#define MPSC_ITERATIONS 100000
+#define MPSC_STACK_SIZE (512 + CONFIG_TEST_EXTRA_STACK_SIZE)
+#define MPSC_THREADS_NUM 4
+
+static struct thread_info mpsc_tinfo[MPSC_THREADS_NUM];
+static struct k_thread mpsc_thread[MPSC_THREADS_NUM];
+static K_THREAD_STACK_ARRAY_DEFINE(mpsc_stack, MPSC_THREADS_NUM, MPSC_STACK_SIZE);
+
+struct mpsc_node {
+	uint32_t id;
+	struct rtio_mpsc_node n;
+};
+
+
+RTIO_SPSC_DECLARE(node_sq, struct mpsc_node, MPSC_FREEQ_SZ);
+
+#define SPSC_INIT(n, sz) RTIO_SPSC_INITIALIZER(sz)
+
+struct rtio_spsc_node_sq node_q[MPSC_THREADS_NUM] = {
+	LISTIFY(MPSC_THREADS_NUM, SPSC_INIT, (,), MPSC_FREEQ_SZ)
+};
+
+static struct rtio_mpsc mpsc_q;
+
+static void mpsc_consumer(void *p1, void *p2, void *p3)
+{
+	struct rtio_mpsc_node *n;
+	struct mpsc_node *nn;
+
+	for (int i = 0; i < (MPSC_ITERATIONS)*(MPSC_THREADS_NUM - 1); i++) {
+		do {
+			n = rtio_mpsc_pop(&mpsc_q);
+			if (n == NULL) {
+				k_yield();
+			}
+		} while (n == NULL);
+
+		zassert_not_equal(n, &mpsc_q.stub, "mpsc should not produce stub");
+
+		nn = CONTAINER_OF(n, struct mpsc_node, n);
+
+		rtio_spsc_acquire(&node_q[nn->id]);
+		rtio_spsc_produce(&node_q[nn->id]);
+	}
+}
+
+static void mpsc_producer(void *p1, void *p2, void *p3)
+{
+	struct mpsc_node *n;
+	uint32_t id = (uint32_t)(uintptr_t)p1;
+
+	for (int i = 0; i < MPSC_ITERATIONS; i++) {
+		do {
+			n = rtio_spsc_consume(&node_q[id]);
+			if (n == NULL) {
+				k_yield();
+			}
+		} while (n == NULL);
+
+		rtio_spsc_release(&node_q[id]);
+		n->id = id;
+		rtio_mpsc_push(&mpsc_q, &n->n);
+	}
+}
+
+/**
+ * @brief Test that the producer and consumer are indeed thread safe
+ *
+ * This can and should be validated on SMP machines where incoherent
+ * memory could cause issues.
+ */
+ZTEST(rtio_mpsc, test_mpsc_threaded)
+{
+	rtio_mpsc_init(&mpsc_q);
+
+	TC_PRINT("setting up mpsc producer free queues\n");
+	/* Setup node free queues */
+	for (int i = 0; i < MPSC_THREADS_NUM; i++) {
+		for (int j = 0; j < MPSC_FREEQ_SZ; j++) {
+			rtio_spsc_acquire(&node_q[i]);
+		}
+		rtio_spsc_produce_all(&node_q[i]);
+	}
+
+	TC_PRINT("starting consumer\n");
+	mpsc_tinfo[0].tid =
+		k_thread_create(&mpsc_thread[0], mpsc_stack[0], MPSC_STACK_SIZE,
+				(k_thread_entry_t)mpsc_consumer,
+				NULL, NULL, NULL,
+				K_PRIO_PREEMPT(5),
+				K_INHERIT_PERMS, K_NO_WAIT);
+
+	for (int i = 1; i < MPSC_THREADS_NUM; i++) {
+		TC_PRINT("starting producer %i\n", i);
+		mpsc_tinfo[i].tid =
+			k_thread_create(&mpsc_thread[i], mpsc_stack[i], MPSC_STACK_SIZE,
+					(k_thread_entry_t)mpsc_producer,
+					(void *)(uintptr_t)i, NULL, NULL,
+					K_PRIO_PREEMPT(5),
+					K_INHERIT_PERMS, K_NO_WAIT);
+	}
+
+	for (int i = 0; i < MPSC_THREADS_NUM; i++) {
+		TC_PRINT("joining mpsc thread %d\n", i);
+		k_thread_join(mpsc_tinfo[i].tid, K_FOREVER);
+	}
+}
+
+ZTEST_SUITE(rtio_mpsc, NULL, NULL, NULL, NULL, NULL);
